+++ lib/packetgen/types/fields.rb
@@ -138,9 +138,6 @@ module PacketGen
              type.instance_methods.include? :from_human
           define << "def #{name}; self[:#{name}].to_human; end"
           define << "def #{name}=(val) self[:#{name}].from_human val; end"
-        elsif type == Types::HTTPHeaders
-          define << "def #{name}; self[:#{name}].data; end\n"
-          define << "def #{name}=(val) self[:#{name}].read val; end"
         else
           define << "def #{name}; self[:#{name}]; end\n"
           define << "def #{name}=(val) self[:#{name}].read val; end"
